requirements.txt
torch>=1.8.0
numpy>=1.19.2
pandas>=1.2.0
matplotlib>=3.3.4
tqdm>=4.56.0
wandb>=0.12.0
bokeh>=2.3.0
pyyaml>=5.4.1
